+++ src/main/java/com/jaceg18/localrealm/App.java
@@ -11,7 +11,7 @@ import java.io.IOException;
 
 public class App extends Application {
 
-    public static final String VERSION = "1.3-SNAPSHOT";
+    public static final String VERSION = "1.3.5-SNAPSHOT";
 
     @Override
     public void start(Stage stage) throws IOException {